+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.4] - 2021-02-11
+### Fixed
+- The ```License``` entity now allows ```null``` for the fields ```price.currency```, ```friendlyName```, and ```vendorSubscriptionId```
+
 ## [0.2.3] - 2021-02-10
 ### Added
 - Customers client (list and create)